Refactor and expand functionality of tools:
Create a Tools base class to hide away some of the boilerplate involved
in parsing command line options and to keep the core functionality of
each tool in a single run() function per tool.

Add a feature to the mkstripes tool which reads colors from stdin.

Add a feature to the getcolors tool which lets the user specify the
depth of the tree used in the color quantization algorithm.

Fix a bug in the getcolors tool which was incorrectly listing the colors
in an image, by pulling the color histogram from the image instead of
the color map. The specific failure case seen was a listing of
"#ffffff"/white in a .png stripes image with no white color.

Simplify the StripesImage class.

Fix some style to uniformly place the const keyword and consistently
initialize class fields in constructors.