Create shell script which generates color wheels.
Write a shell script which generates a collection of two-dimensional
subsets of color spaces as .png images using ImageMagick commands.

Example: for the RGB color space, create a collection of images
displaying the subset of the RGB color space where one channel (such as
the red component) is held constant. More specific example: generate the
subset of the RGB colorspace where the R component of every pixel is
held constant at 50% of its maximum value, and export that subset to a
.png file.

The script also generates each color space subset at different scales,